+++ pandas/tests/indexes/datetimes/test_timezones.py
@@ -596,25 +596,6 @@ class TestDatetimeIndexTimezones:
         assert idx[0].tzinfo is not None
 
     @pytest.mark.parametrize("tzstr", ["US/Eastern", "dateutil/US/Eastern"])
-    def test_dti_astype_asobject_tzinfos(self, tzstr):
-        # GH#1345
-
-        # dates around a dst transition
-        rng = date_range("2/13/2010", "5/6/2010", tz=tzstr)
-
-        objs = rng.astype(object)
-        for i, x in enumerate(objs):
-            exval = rng[i]
-            assert x == exval
-            assert x.tzinfo == exval.tzinfo
-
-        objs = rng.astype(object)
-        for i, x in enumerate(objs):
-            exval = rng[i]
-            assert x == exval
-            assert x.tzinfo == exval.tzinfo
-
-    @pytest.mark.parametrize("tzstr", ["US/Eastern", "dateutil/US/Eastern"])
     def test_dti_with_timezone_repr(self, tzstr):
         rng = date_range("4/13/2010", "5/6/2010")
